+++ src/app/components/TodoList/TodoList.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CircularProgress from 'material-ui/CircularProgress';
 
 import { todoShape, Todo } from '../../models';
 import TodoComponent from '../Todo';
@@ -9,41 +8,27 @@ const styles = require('./TodoList.scss');
 
 export interface TodoListProps {
   todos: Todo[];
-  onTodoClick: (id: number) => void;
-  loading?: boolean;
+  onTodoClick: (id: string) => void;
 }
 
-const TodoList: React.SFC<TodoListProps> = ({ todos, onTodoClick, loading }) => {
+const TodoList: React.SFC<TodoListProps> = ({ todos, onTodoClick }) => {
   return (
-    <div>
-      {loading ? (
-        <div className={styles.progressContainer}>
-          <CircularProgress />
-        </div>
-      ) : (
-        <ul className={styles.todoList}>
-          {todos.map(todo => (
-            <TodoComponent
-              key={todo.id}
-              onClick={() => onTodoClick(todo.id)}
-              completed={todo.completed}
-              text={todo.text}
-            />
-          ))}
-        </ul>
-      )}
-    </div>
+    <ul className={styles.todoList}>
+      {todos.map(todo => (
+        <TodoComponent
+          key={todo.id}
+          onClick={() => onTodoClick(todo.id)}
+          completed={todo.completed}
+          text={todo.text}
+        />
+      ))}
+    </ul>
   );
 };
 
 TodoList.propTypes = {
   todos: PropTypes.arrayOf(todoShape).isRequired,
   onTodoClick: PropTypes.func.isRequired,
-  loading: PropTypes.bool,
-};
-
-TodoList.defaultProps = {
-  loading: false,
 };
 
 export default TodoList;