+++ src/app/containers/AddTodo/AddTodo.tsx
@@ -1,26 +1,27 @@
 import { Dispatch } from 'redux';
-import { connect } from 'react-redux';
+import { connect, MapDispatchToProps } from 'react-redux';
 
 import { State } from '../../reducers';
 import { addTodo, fetchTodos } from '../../actions';
-import { getPending, getError } from '../../selectors';
 import AddTodo from '../../components/AddTodo';
 
-const mapStateToProps = (state: State) => ({
-  disabled: getPending(state) || !!getError(state),
-});
+export interface AddTodoDispatchProps {
+  onSubmit: (todo: string) => void;
+  onInit: () => void;
+}
 
-const mapDispatchToProps = (dispatch: Dispatch<State>) => ({
-  onSubmit: (todo: string) => {
-    dispatch(addTodo(todo));
-  },
-  onInit: () => {
-    dispatch(fetchTodos());
-  },
-});
+const mapDispatchToProps: MapDispatchToProps<AddTodoDispatchProps, {}> =
+  (dispatch: Dispatch<State>) => ({
+    onSubmit: (todo: string) => {
+      dispatch(addTodo(todo));
+    },
+    onInit: () => {
+      dispatch(fetchTodos());
+    },
+  });
 
-const AddTodoContainer = connect<any, any, any>(
-  mapStateToProps,
+const AddTodoContainer = connect<{}, AddTodoDispatchProps, {}>(
+  undefined,
   mapDispatchToProps,
 )(AddTodo);