+++ src/app/epics/todos.ts
@@ -1,14 +1,15 @@
 import { Observable } from 'rxjs/Observable';
+import { ActionsObservable } from 'redux-observable';
 import { Store } from 'redux';
 
 import { Action, fetchTodosFulfilled, fetchTodosPending, fetchTodosRejected } from '../actions';
 import { getTodos } from '../api';
-import { FETCH_TODOS } from '../constants';
+import { ActionTypes } from '../constants';
 import { State } from '../reducers';
 
-export const fetchTodosEpic = (action$: Observable<Action<any, any>>, store: Store<State>) =>
+export const fetchTodosEpic = (action$: ActionsObservable<Action<any, any>>, store: Store<State>) =>
   action$
-    .filter(action => action.type === FETCH_TODOS)
+    .ofType(ActionTypes.FETCH_TODOS)
     .do(() => store.dispatch(fetchTodosPending()))
     .mergeMap(() =>
       getTodos()