+++ src/app/models/index.ts
@@ -1,13 +1,26 @@
 import PropTypes from 'prop-types';
+import { schema } from 'normalizr';
 
 export interface Todo {
-  id: number;
+  id: string;
   completed: boolean;
   text: string;
 }
 
 export const todoShape = PropTypes.shape({
-  id: PropTypes.number.isRequired,
+  id: PropTypes.string.isRequired,
   completed: PropTypes.bool.isRequired,
   text: PropTypes.string.isRequired,
 });
+
+export const todoSchema = new schema.Entity('todos');
+export const todoListSchema = [todoSchema];
+
+export interface NormalizedTodoList {
+  entities: {
+    todos: {
+      [id: string]: Todo;
+    };
+  };
+  result: string[];
+}