+++ src/app/reducers/index.ts
@@ -1,20 +1,13 @@
 import { combineReducers } from 'redux';
 
-import { todosReducer as todos } from './todos';
-import { appStatusReducer as appStatus, AppStatusState } from './appStatus';
-import { Todo } from '../models';
+import { todosReducer as todos, TodosState } from './todos';
 
-/**
- * Root reducer
- */
 const rootReducer = combineReducers<State>({
   todos,
-  appStatus,
 });
 
 export interface State {
-  todos?: Todo[];
-  appStatus?: AppStatusState;
+  todos?: TodosState;
 }
 
 export default rootReducer;